Update dependencies (#1)
* update dependencies

* add names + fix gizmo rendering + fix initial state of visibility explorer

* add vscode feature conf + launch

+++ src/layers.rs
@@ -1,1 +1,1 @@
-pub const GIZMO_LAYER: u8 = 1;
+pub const GIZMO_LAYER: usize = 1;
